refactor(voice/capture): extract RingMixin (4 methods, ~190 LOC) [T1.4 step 7]
Master mission Phase 1 / T1.4 — seventh stage. Builds on step 6
(`c433cc0`, EpochMixin) by landing the second mixin. The pattern
is now proven across two extractions; subsequent steps
(RestartMixin, LoopMixin) follow the same blueprint.

What lands:

  * New module ``capture/_ring.py`` with class ``RingMixin``:
    - Five host-class state declarations
      (``_ring_buffer``, ``_ring_capacity``, ``_ring_write_index``,
      ``_ring_state``, ``_tuning``) — mypy-strict-compatible
      attribute annotations.
    - Four methods:
      * ``_allocate_ring_buffer`` — allocate / resize the 16 kHz
        mono int16 ring; bumps the epoch on every (re)allocation.
      * ``_ring_write`` — synchronous append-with-wrap, called
        between awaits in (eventual) ``LoopMixin._consume_loop``.
      * ``tap_recent_frames`` — async public API; returns a fresh
        copy of the most recent ``duration_s`` of audio.
      * ``tap_frames_since_mark`` — async public API; polls the
        ring state until ``min_samples`` accumulate post-``mark``
        or ``max_wait_s`` elapses (probe-window-contamination
        fix per v1.3 §4.2 L4-B).
    Total ~250 LOC including module + class docstrings.
  * ``_capture_task.py``:
    - ``class AudioCaptureTask(EpochMixin)`` →
      ``class AudioCaptureTask(EpochMixin, RingMixin)``.
    - The 4 ring methods removed (~190 LOC).
    - Import ``RingMixin`` from ``capture._ring``.
  * ``capture/__init__.py`` re-exports ``RingMixin`` alongside
    ``EpochMixin`` from step 6.

External callers + Protocol contract:

  * ``CaptureTaskProto`` in ``health/contract/_probe_result.py``
    expects ``tap_recent_frames`` + ``tap_frames_since_mark`` on
    the AudioCaptureTask surface — MRO resolution preserves the
    Protocol satisfaction.
  * ``capture_integrity.py`` calls ``self._capture_task.<method>``
    on AudioCaptureTask instances — MRO resolution.
  * Tests at ``test_capture_task.py:1320, 1353, 1382`` use the
    methods directly via the AudioCaptureTask instance — no
    migration needed.

T1.4 cumulative impact (steps 1-7):

  * ``_capture_task.py``: 2785 → 1976 LOC (~809 LOC extracted,
    ~29 % of the split done; cumulative drop ~29 %).
  * Seven sub-modules under ``voice/capture/``:
    ``_restart`` (verdict types), ``_exceptions``, ``_contention``,
    ``_helpers``, ``_constants``, ``_epoch`` (1st mixin),
    ``_ring`` (2nd mixin).
  * Pattern proven on 2 mixins. Steps 8-10 (RestartMixin ~970 LOC,
    LoopMixin ~700 LOC, final cleanup) follow per the plan doc
    at ``docs-internal/T1.4-step-6-mixin-surgery-plan.md``.

Validation:
  * ruff check + format clean (auto-fix unquoted forward
    references in ``capture/_ring.py``; resorted imports in
    ``_capture_task.py`` and ``capture/__init__.py``)
  * mypy strict: 0 issues across 426 source files (+1 source file
    for ``capture/_ring.py``)
  * bandit clean
  * pytest full suite: 11894 passed, 25 skipped, 57 deselected
    (zero delta — pure structural extraction; MRO resolution
    keeps every external caller working unchanged)

Mission: docs-internal/missions/MISSION-voice-final-skype-grade-2026.md §Phase 1.T1.4 (step 7)
